config: SSR Migration is Done.
- resolved Hydration mismatch Error
- resolved `No routes matched location "test"` by react-router-dom Error
