+++ src/components/Test.tsx
@@ -1,17 +1,6 @@
-import { useEffect, useState } from "react";
 import styles from "../app.module.scss";
 
 const Test = () => {
-  const [isClient, setIsClient] = useState(false);
-
-  useEffect(() => {
-    setIsClient(true);
-  }, []);
-
-  if (!isClient) {
-    return null;
-  }
-
   return (
     <div>
       <span className={styles.debug}>test</span>
